refactored code, testing elasticache
