+++ src/app/utils/aws/dynamodb.py
